+++ lib/rendering/render-page.js
@@ -7,7 +7,7 @@ import { getPageInfo } from './helpers/site-map';
 export default function renderPage(templatePath, nunjucksEnvironment) {
   try {
     const data = frontMatter(fs.readFileSync(templatePath, { encoding: 'utf8' }));
-    const layout = !!data.attributes.layout ? `prototypes/${data.attributes.layout}` : 'page-templates/_template';
+    const layout = !!data.attributes.layout ? `prototypes/${data.attributes.layout}` : 'layout/_template';
     const body = data.body;
 
     let template;
