+++ package.json
@@ -86,6 +86,7 @@
     "postcss-url": "^8.0.0",
     "prettier": "^1.16.4",
     "pretty": "^2.0.0",
+    "querystring-helpers": "^1.0.2",
     "remark": "^10.0.1",
     "remark-cli": "^6.0.1",
     "remark-lint": "^6.0.4",
@@ -97,8 +98,6 @@
   },
   "dependencies": {
     "@ons/design-system": "^21.1.0",
-    "@ons/design-system/0.5.0": "npm:@ons/design-system@0.5.0",
-    "@ons/design-system/44.0.0": "npm:@ons/design-system@44.0.0",
-    "querystring-helpers": "^1.0.2"
+    "@ons/design-system/45.1.3": "npm:@ons/design-system@45.1.3"
   }
 }
