+++ src/helpers/previous-link.js
@@ -1,11 +1,26 @@
-import domready from 'helpers/domready';
+import domready from './domready';
 
 domready(() => {
-  const previousURL = new URLSearchParams(window.location.search).get('previous');
+  let previousURL;
+  const queryStringPreviousURL = new URLSearchParams(window.location.search).get('previous');
+
+  if (queryStringPreviousURL) {
+    previousURL = queryStringPreviousURL;
+  } else {
+    const savedData = sessionStorage.getItem(window.location.pathname);
+
+    if (savedData) {
+      previousURL = JSON.parse(savedData).previousURL;
+    }
+  }
 
   if (previousURL) {
-    const previousLinks = [...document.querySelectorAll('.js-previous')];
+    const previousLinks = [...document.querySelectorAll('.ons-js-previous')];
 
-    previousLinks.forEach(link => link.setAttribute('href', previousURL));
+    previousLinks.forEach(link => {
+      const originalPreviousURL = link.getAttribute('href');
+      link.setAttribute('href', previousURL);
+      link.setAttribute('data-original-href', originalPreviousURL);
+    });
   }
 });
