+++ src/helpers/summary-manager.js
