+++ src/index.njk
@@ -1,9 +1,9 @@
 ---
-version: 0.5.0
+version: 45.1.3
 ---
 {% from "components/lists/_macro.njk" import onsList %}
 
-{% set page = { 
+{% set pageConfig = { 
     "title": "ONS Prototypes",
     "footer": {
         "OGLLink": {
@@ -14,7 +14,7 @@ version: 0.5.0
         }
     }
 } %}
-{% block content %}
+{% block main %}
     {% set groups = helpers.navigationHelper({ 
         "pages": prototypeListing,
         "ignorePaths": ["/"]
@@ -22,10 +22,10 @@ version: 0.5.0
     {% for group in groups %}
         <section>
             <h1 class="u-fs-m">{{ group.title }}</h1>
-            {{ 
+            {{
                 onsList({
-                    "classes": "list--bare",
-                    "items": group.items
+                    "classes": "ons-list--bare",
+                    "itemsList": group.items
                 })
             }}
         </section>
