--- src/prototypes/ds-features-nov2021/components/summary/_summary.scss
@@ -1,183 +1,0 @@
-$summary-row-spacing: 1rem;
-$summary-col-spacing: 1rem;
-$hub-row-spacing: 1.3rem;
-
-.ons-summary {
-  &__items {
-    border-collapse: collapse;
-    border-spacing: 0;
-    width: 100%;
-
-    + .ons-summary__group-title {
-      margin-top: 1.5rem;
-    }
-  }
-
-  &__item {
-    line-height: 1.4;
-
-    &:not(:last-child),
-    &:nth-of-type(1) {
-      border-bottom: 1px solid $color-borders;
-    }
-
-    &--total {
-      @extend .ons-u-fs-m;
-
-      border-width: 2px;
-      font-weight: 700;
-    }
-
-    &--error {
-      background: $color-errors-tint;
-      border-left: 8px solid $color-errors;
-    }
-  }
-
-  &__row-title {
-    padding: $summary-row-spacing 0;
-    text-align: left;
-  }
-
-  &__item-title,
-  &__values,
-  &__actions {
-    hyphens: manual;
-    overflow-wrap: break-word;
-    padding: 0 0 $summary-row-spacing;
-    vertical-align: top;
-    word-wrap: break-word;
-  }
-
-  &__item-title {
-    padding-top: $summary-row-spacing;
-    position: relative;
-    &--text {
-      display: block;
-      overflow: hidden;
-      padding-left: 2rem;
-    }
-    &-icon {
-      left: 0;
-      position: absolute;
-      text-align: center;
-      &--check .ons-svg-icon {
-        fill: $color-leaf-green !important;
-      }
-    }
-  }
-
-  &__actions {
-    white-space: nowrap;
-  }
-
-  &__spacer {
-    background: $color-black;
-    display: inline-block;
-    height: 1rem;
-    margin: 0 0.25rem;
-    vertical-align: middle;
-    width: 1px;
-  }
-
-  // Item Modifiers
-  &__item--total & {
-    &__values {
-      @extend .ons-u-fs-m;
-    }
-  }
-
-  &__item--error & {
-    &__row-title {
-      color: $color-errors;
-      font-weight: 700;
-      padding: $summary-row-spacing $summary-col-spacing;
-    }
-
-    &__item-title,
-    &__values,
-    &__actions {
-      padding-left: $summary-col-spacing;
-      padding-right: $summary-col-spacing;
-
-      @include mq('s') {
-        padding-left: $summary-col-spacing / 2;
-        padding-right: $summary-col-spacing / 2;
-
-        &:first-child {
-          padding-left: $summary-col-spacing;
-        }
-
-        &:last-child {
-          padding-right: $summary-col-spacing;
-        }
-      }
-    }
-  }
-
-  // Modifiers
-  &--hub & {
-    &__actions {
-      padding: 0 0 $hub-row-spacing;
-    }
-
-    &__item-title {
-      @extend .ons-u-fs-r--b;
-
-      padding-top: $hub-row-spacing;
-    }
-  }
-
-  &:not(&--hub) & {
-    &__values {
-      @extend .ons-u-fs-r--b;
-    }
-  }
-
-  // Breakpoints
-  @include mq(xxs, s, none, '<') {
-    &__item-title,
-    &__values,
-    &__actions {
-      display: block;
-    }
-
-    &--hub & {
-      &__values {
-        display: none;
-      }
-    }
-  }
-
-  @include mq(s) {
-    &__item-title,
-    &__values,
-    &__actions {
-      padding-top: $summary-row-spacing;
-      vertical-align: top;
-
-      &:not(:last-child) {
-        padding-right: $summary-col-spacing;
-      }
-    }
-
-    &__actions {
-      text-align: right;
-    }
-
-    &__row--has-values & {
-      &__item-title,
-      &__values {
-        width: 50%;
-      }
-    }
-
-    &--hub & {
-      &__item-title,
-      &__values,
-      &__actions {
-        padding-top: $hub-row-spacing;
-      }
-    }
-  }
-}
