--- src/prototypes/ds-features-nov2021/enter-access-code.njk
@@ -1,141 +1,0 @@
----
-version: 44.0.0
-layout: ds-features-nov2021/views/_master
----
-{% from "components/access-code/_macro.njk" import onsUACInput %}
-{% from "components/button/_macro.njk" import onsButton %}
-{% from "components/panel/_macro.njk" import onsPanel %}
-{% from "components/collapsible/_macro.njk" import onsCollapsible %}
-{% from "components/lists/_macro.njk" import onsList %}
-
-{% set backLink = "Back" %}
-{% set prevPage = "index" %}
-{% block main %}
-    {% call onsPanel({
-            "id": "js-error-summary",
-            "classes": "ons-u-d-no",
-            "type": "error",
-            "title": "There is a problem with this page"
-        })
-    %}
-        {{
-            onsList({
-                "classes": "js-error-summary-items",
-                "element": "ol",
-                "itemsList": [
-                    {
-                        "text": "",
-                        "url": "#uac-error",
-                        "variants": "inPageLink"
-                    }
-                ]
-            })
-        }}
-    {% endcall %}
-
-    <h1 class="ons-u-mt-l">Start study</h1>
-
-    {{ onsUACInput({
-        "id": "uac",
-        "name": "uac",
-        "postTextboxLinkText": "Where to find your access code",
-        "postTextboxLinkUrl": "#0",
-        "label": {
-            "text": "Enter your 16-character access code",
-            "description": "Keep this code safe. You will need to enter it every time you access your study"
-        },
-        "securityMessage": "Your personal information is protected by law and will be kept confidential"
-    }) }}
-
-    {{ onsButton({
-        "id": "js-submit",
-        "text": "Access study",
-        "classes": "ons-u-mb-xl"
-    }) }}
-
-    <h2>If you don’t have an access code</h2>
-    <p>
-        You can <a href="#0">request a new access code</a> to start a new study if you have lost or not received an access code. This can be sent to you by text or post.
-    </p>
-
-    {% set content %}
-        <p>If you need to answer separately from the people you live with, you can <a href='#0'>request an individual access code</a> to start a separate study.</p>
-        {% call onsPanel({
-                type: "warn"
-            })
-        %}
-            Someone in your household must still complete a study using a household access code
-        {% endcall %}
-    {% endset %}
-
-    {% call onsCollapsible({
-            id: "collapsible",
-            title: "Need to answer separately from your household?",
-            titleTag: "h2",
-            button: {
-                "close": "Hide this",
-                "contextSuffix": "content"
-            }
-        })
-    %}
-        {{ content | safe }}
-    {% endcall %}
-
-    <script>
-        const uacInput = document.getElementById('uac');
-        const submitBtn = document.getElementById('js-submit');
-        const errorSummary = document.getElementById('js-error-summary');
-        const errorSummaryLink = document.querySelector('.js-error-summary-items').getElementsByTagName('a')[0];
-        const errorPanel = document.querySelector('.ons-panel--info');
-        const errorPanelBody = document.querySelector('.ons-panel--no-title').querySelector('.ons-panel__body');
-        const errorPanelMessagePara = document.createElement('p');
-        const errorPanelMessageStrong = document.createElement('strong');
-        const errorPanelMessageText = document.createTextNode("This is an error");
-        const inputField = document.querySelector('.ons-question__answer');
-
-        document.addEventListener("keyup", function(e) {
-            if (e.keyCode === 13) {
-                e.preventDefault();
-                submitBtn.click();
-            }
-        })
-
-        submitBtn.addEventListener("click", function() {
-            if (uacInput.value === 'A1C9 F0L1 I7D4 V3O0') {
-                window.location = "accordions.html";
-            }
-            else {
-                errorSummary.classList.remove('ons-u-d-no');
-                errorSummaryLink.classList.add('ons-js-inpagelink');
-                errorSummary.focus();
-                errorPanel.classList.replace('ons-panel--info', 'ons-panel--error');
-                errorSummaryLink.setAttribute('id', 'uac-error');
-                errorPanelMessagePara.classList.add('ons-panel__error');
-                errorPanelMessagePara.appendChild(errorPanelMessageStrong);
-                errorPanelMessageStrong.appendChild(errorPanelMessageText);
-                errorPanelBody.insertBefore(errorPanelMessagePara, inputField);
-                errorPanelMessagePara.classList.add('ons-panel__error');
-
-                errorSummaryLink.addEventListener("click", function(e) {
-                    e.preventDefault();
-                    errorPanel.scrollIntoView();
-                    uacInput.focus();
-                });
-
-                if (!uacInput.value.length){
-                    errorSummaryLink.innerHTML = 'Enter an access code';
-                    errorPanelMessageStrong.innerHTML = 'Enter an access code';
-                }
-                else if (uacInput.value.length < 19) {
-                    errorSummaryLink.innerHTML = 'Enter a 16-character access code';
-                    errorPanelMessageStrong.innerHTML = 'Enter a 16-character access code';
-                }
-                else if (uacInput.value.length == 19) {
-                    errorSummaryLink.innerHTML = 'Access code not recognised. Enter the code again.';
-                    errorPanelMessageStrong.innerHTML = 'Access code not recognised. Enter the code again.';
-                }
-            }
-        });
-
-    </script>
-{% endblock %}
