--- src/prototypes/ds-features-nov2021/name.njk
@@ -1,63 +1,0 @@
----
-version: 44.0.0
-layout: ds-features-nov2021/views/_master
----
-{% from "components/question/_macro.njk" import onsQuestion %}
-{% from "components/fieldset/_macro.njk" import onsFieldset %}
-{% from "components/input/_macro.njk" import onsInput %}
-{% from "components/button/_macro.njk" import onsButton %}
-
-{% set questionTitle = "What is your name?" %}
-{% set backLink = "Back" %}
-{% set prevPage = "accordions" %}
-{% set form =  {
-    "method": "GET",
-    "attributes": {
-        "action":  pageInfo.rootPath + "/date.html"
-    }
-} %}
-
-{% block main %}
-    {% call onsQuestion({
-        "title": questionTitle,
-        "legendIsQuestionTitle": true
-    }) %}
-        {% call onsFieldset({
-                "legend": questionTitle,
-                "dontWrap": true
-        }) %}
-            {{
-                onsInput({
-                    "id": "first-name",
-                    "label": {
-                        "text": "First name"
-                    }
-                })
-            }}
-            {{
-                onsInput({
-                    "id": "middle-name",
-                    "label": {
-                        "text": "Middle names"
-                    }
-                })
-            }}
-            {{
-                onsInput({
-                    "id": "last-name",
-                    "label": {
-                        "text": "Last name"
-                    }
-                })
-            }}
-        {% endcall %}
-
-
-    {% endcall %}
-
-    {{ onsButton({
-        "text": "Save and continue",
-        "classes": "ons-u-mt-xl"
-    }) }}
-
-{% endblock %}
