--- src/prototypes/ds-features-nov2021/radios.njk
@@ -1,59 +1,0 @@
----
-version: 44.0.0
-layout: ds-features-nov2021/views/_master
----
-{% from "components/question/_macro.njk" import onsQuestion %}
-{% from "components/fieldset/_macro.njk" import onsFieldset %}
-{% from "components/input/_macro.njk" import onsInput %}
-{% from "components/button/_macro.njk" import onsButton %}
-{% from "components/radios/_macro.njk" import onsRadios %}
-
-{% set questionTitle = "Do you have a passport?" %}
-{% set prevPage = "duration" %}
-{% set form =  {
-    "method": "GET",
-    "attributes": {
-        "action":  pageInfo.rootPath + "/autosuggest-countries.html"
-    }
-} %}
-
-{% block main %}
-    {% call onsQuestion({
-        "title": questionTitle,
-        "legendIsQuestionTitle": true
-    }) %}
-
-    {{
-        onsRadios({
-            "name": "passport",
-            "dontWrap": true,
-            "radios": [
-                {
-                    "id": "yes",
-                    "label": {
-                        "text": "Yes"
-                    },
-                    "value": "Yes"
-                },
-                {
-                    "id": "no",
-                    "label": {
-                        "text": "No"
-                    },
-                    "value": "No",
-                    "attributes": {
-                        "data-action-url":  pageInfo.rootPath + "/mutually-exclusive.html"
-                    }
-                }
-            ]
-        })
-    }}
-
-    {% endcall %}
-
-    {{ onsButton({
-        "text": "Save and continue",
-        "classes": "ons-u-mt-xl"
-    }) }}
-
-{% endblock %}
