--- src/prototypes/ds-features-nov2021/summary.njk
@@ -1,50 +1,0 @@
----
-group: Design system
-title: Features test - 22 Nov 2021
-version: 44.0.0
-layout: ds-features-nov2021/views/_master
----
-{% from "components/button/_macro.njk" import onsButton %}
-{% from "components/panel/_macro.njk" import onsPanel %}
-
-{% set prevPage = "mutually-exclusive-quals" %}
-{% set backLink = "Back" %}
-{% set form =  {
-    "method": "GET",
-    "attributes": {
-        "action":  pageInfo.rootPath + "/confirmation.html"
-    }
-} %}
-
-{% block main %}
-    <h1>Submit study</h1>
-    {% call onsPanel({
-            "type": "warn"
-        })
-    %}
-        You must submit this study to complete it
-    {% endcall %}
-
-    <div class="ons-summary">
-        <div class="ons-summary__group">
-            <h2 class="ons-summary__group-title">Turnover</h2>
-            <table id="turnover" class="ons-summary__items">
-                <thead class="ons-u-vh">
-                    <tr>
-                    <th>Question</th>
-                    <th>Answer given</th>
-                    <th>Change answer</th>
-                    </tr>
-                </thead>
-                <table class="js-summary"></table>
-            </table>
-        </div>
-    </div>
-
-    <p class="ons-u-mt-l">By submitting this study you are confirming that, to the best of your knowledge and belief, the details provided are correct.</p>
-
-    {{ onsButton({
-        "text": 'Submit study'
-        })
-    }}
-{% endblock %}
