--- src/prototypes/ds-features-nov2021/views/partials/_summary.njk
@@ -1,3 +1,0 @@
-{% from "!nunjucks-loader!../../components/summary/_macro.njk" import onsSummary %}
-
-{{ onsSummary(params) }}
