+++ src/prototypes/example/index.njk
@@ -1,12 +1,12 @@
 ---
 group: Examples
 title: Example
-version: 0.5.0
+version: 45.1.3
 ---
 {% from "components/input/_macro.njk" import onsInput %}
 {% from "components/button/_macro.njk" import onsButton %}
 
-{% set page = { 
+{% set pageConfig = { 
     "title": "Example Prototype",
     "footer": {
         "OGLLink": {
@@ -22,7 +22,7 @@ version: 0.5.0
     <link rel="stylesheet" href="/prototypes/example/styles.css" />
 {% endblock %}
 
-{% block content %}
+{% block main %}
     <form class="u-mt-m">
         <h1>What is your name?</h1>
         {{
@@ -30,7 +30,7 @@ version: 0.5.0
                 "id": "first-name",
                 "name": "first-name",
                 "label": {
-                    "text": "First or given name"
+                    "text": "First name"
                 }
             })
         }}
@@ -39,19 +39,20 @@ version: 0.5.0
                 "id": "last-name",
                 "name": "last-name",
                 "label": {
-                    "text": "Surname or family name"
+                    "text": "Surname"
                 }
             })
         }}
         {{
             onsButton({
                 "text": "Save and continue",
-                "classes": "u-mt-l"
+                "classes": "ons-u-mt-l"
             })
         }}
     </form>
 {% endblock %}
 
 {% block bodyEnd %}
+    <script src="/helpers/index.js"></script>
     <script src="/prototypes/example/index.js"></script>
 {% endblock %}
