+++ src/prototypes/example/styles.scss
@@ -1,8 +1,13 @@
-@import '../../../node_modules/@ons/design-system/44.0.0/scss/vars/_index.scss';
-@import '../../../node_modules/@ons/design-system/44.0.0/scss/helpers/_index.scss';
-@import '../../../node_modules/@ons/design-system/44.0.0/scss/base/_index.scss';
-@import '../../../node_modules/@ons/design-system/44.0.0/scss/utilities/_index.scss';
+/***********************************************************************
+Update the version in the path below to the version being used
+***********************************************************************/
+@import '../../../node_modules/@ons/design-system/45.1.3/scss/vars/_index.scss';
+@import '../../../node_modules/@ons/design-system/45.1.3/scss/helpers/_index.scss';
+@import '../../../node_modules/@ons/design-system/45.1.3/scss/utilities/_index.scss';
 
+/***********************************************************************
+Do not remove. Required for dynamic generation of survey summary page
+***********************************************************************/
 .ons-summary__item:not(:last-child) {
   border-bottom: 1px solid $color-borders !important;
 }
@@ -12,3 +17,5 @@
   border-spacing: 0;
   width: 100%;
 }
+/***********************************************************************/
+/***********************************************************************/
