+++ src/prototypes/example/views/_master.njk
@@ -1,9 +1,6 @@
-{% extends "page-templates/_template.njk" %}
+{% extends "layout/_template.njk" %}
 
 {% set pageConfig = {
-    "cdn": {
-      "version": '44.0.0'
-    },
     "header": {
         "logoHref": '/',
         "titleLogoHref": pageInfo.rootPath + '#0'
